Can search using regular expressions.

+++ input.cpp
@@ -403,7 +403,7 @@ InputRecord SelectInput(DWORD timeout)
 
 bool ReadInput(StrW& out, History hindex, DWORD max_width, std::optional<std::function<int32(const InputRecord&)>> input_callback)
 {
-    static std::vector<StrW> s_histories[History::MAX];
+    static std::vector<StrW> s_histories[size_t(History::MAX)];
     std::vector<StrW>* const history = (size_t(hindex) < _countof(s_histories)) ? &s_histories[size_t(hindex)] : nullptr;
 
     out.Clear();
